requirements.txt
flask==3.0.0
requests==2.31.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2
google-auth-oauthlib
python-dotenv==1.0.0
pytz==2023.3
APScheduler==3.10.4
gunicorn==21.2.0
uvicorn

